Resgroup catalog changes
Change initial contents in pg_resgroupcapability:
* Remove memory_redzone_limit;
* Add memory_shared_quota, memory_spill_ratio;

Change resgroup concurrency range to [1, 'max_connections']:
* Original range is [0, 'max_connections'], and -1 means unlimited.
* Now the range is [1, 'max_connections'], and -1 is not supported.

Change resgroup limit type from float to int.

Changed below resgroup resource limit types from float to int percentage value:
* cpu_rate_limit;
* memory_limit;
* memory_shared_quota;
* memory_spill_ratio;

+++ src/backend/utils/resgroup/resgroup-ops-dummy.c
@@ -118,10 +118,10 @@ ResGroupOps_UnLockGroup(Oid group, int fd)
 /*
  * Set the cpu rate limit for the OS group.
  *
- * cpu_rate_limit should be within (0.0, 1.0].
+ * cpu_rate_limit should be within [0, 100].
  */
 void
-ResGroupOps_SetCpuRateLimit(Oid group, float cpu_rate_limit)
+ResGroupOps_SetCpuRateLimit(Oid group, int cpu_rate_limit)
 {
 	unsupported_system();
 }
