+++ src/include/utils/resgroup-ops.h
@@ -24,7 +24,7 @@ extern void ResGroupOps_DestroyGroup(Oid group);
 extern void ResGroupOps_AssignGroup(Oid group, int pid);
 extern int ResGroupOps_LockGroup(Oid group, bool block);
 extern void ResGroupOps_UnLockGroup(Oid group, int fd);
-extern void ResGroupOps_SetCpuRateLimit(Oid group, float cpu_rate_limit);
+extern void ResGroupOps_SetCpuRateLimit(Oid group, int cpu_rate_limit);
 extern int64 ResGroupOps_GetCpuUsage(Oid group);
 extern int ResGroupOps_GetCpuCores(void);
 extern int ResGroupOps_GetTotalMemory(void);
